refactor: change path services, fix css properties
