+++ src/api/getWords.ts
@@ -15,25 +15,6 @@ export interface IWords {
     textExampleTranslate: string
 }
 
-export interface IWords {
-  id: string,
-  group: number,
-  page: number,
-  word: string,
-  image: string,
-  audio: string,
-  audioMeaning: string,
-  audioExample: string,
-  textMeaning: string,
-  textExample: string,
-  transcription: string,
-  wordTranslate: string,
-  textMeaningTranslate: string,
-  textExampleTranslate: string
-}
-
-
-
 export const getWords = async(page: number, group: number) => {
   const response = await fetch(`https://rs-lang-learn.herokuapp.com/words?page=${page}&group=${group}`, {
     headers: {
