+++ src/pages/sprintPage/sprintPage.css
@@ -260,6 +260,7 @@
 
 .mini-audio{
   margin-left: 47px;
+  cursor: pointer;
 }
 
 .point-sum{
