+++ src/services/renderFormAudioCall.ts
@@ -1,0 +1,57 @@
+export const renderSvgExit = `
+<a href="#game" class="svg-exit"></a>
+`
+export const renderProgressGame = `
+<div class="answer-point">
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+  <div class="heart"></div>
+</div>
+`
+export const renderGame = `
+<div class="img-volume-position">
+  <button class="svg-volume"></button>
+</div> 
+<div class="answers">
+  <div class="level-1">
+    <button class="answer-item answer-item-1">
+      <span class="answer-1">1</span>
+    </button>
+  </div>
+  <div class="level-1">
+    <button class="answer-item answer-item-2">
+      <span class="answer-2">2</span>
+    </button>
+  </div>
+  <div class="level-1">
+    <button class="answer-item answer-item-3">
+      <span class="answer-3">3</span>
+    </button>
+  </div>
+  <div class="level-1">
+    <button class="answer-item answer-item-4">
+      <span class="answer-4">4</span>
+    </button>
+  </div>
+  <div class="level-1">
+    <button class="answer-item answer-item-5">
+      <span class="answer-5">5</span>
+    </button>
+  </div>
+</div>
+`
+export const renderNextButton = `
+<button class="next-word">Следующий</button>
+`
